Don't call it "persist model to kernel" any more.
Improve the naming of these menu entries in the Model Editor. "Upload to
kernel" is more intuitive / easier to understand than "persist to
kernel". While here, call it "download from kernel" instead of "load
from kernel" for the sake of consistency.

Also adjust the wording in the User's Guide as well as the names of a
couple of related classes and methods accordingly.

While here, remove a couple of superfluous members from FileMenu.

Reviewed-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>