Set related vehicle and order for peripheral jobs
Apparently, the related vehicle and transport order of a peripheral job
were never actually set, so peripheral comm adapters could not access
this information when processing a job. With this change, both are
properly set.

Also ensure that, when providing the peripheral job object to the comm
adapter for processing, it is the latest version of the object.

Approved-by: Leonard Schüngel <leonard.schuengel@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>