Add support for manual/forced rerouting
Rerouting can now be triggered explicitly via the API, and can
optionally be forced to allow a vehicle to be rerouted from a position
we had not sent it to. This may be necessary in situtations in which a
vehicle encounters an obstacle on its route that cannot be removed and
also cannot be routed around any more, i.e. when the vehicle finds
itself in a one-way dead-end. The vehicle can then be moved to a
different position (manually) and rerouted from there.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>