Reuse Docker images for code quality CI jobs
Add a Dockerfile for the code quality job in the GitLab CI pipeline, and
add a variable that can be used to trigger (re-)building it. This way,
the code quality job doesn't have to download and install all
dependencies with every build, which makes builds a bit faster and more
sustainable.

Approved-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>