Only update the kernel if the pause state changed
The VehiclePausedAdapter reacts to changes of the vehicle pause state
and sends those changes to the kernel. This can cause a feedback loop
where the pause state is changed multiple times to the same value.
To avoid the feedback we only update the kernel if the pause state
is different.

Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>