Eliminate JavaDoc author tags
There are no simple rules for how to maintain author tags, e.g. when
modifying existing code, and these tags do not really serve a purpose,
so we get rid of them here. The only exceptions where tags are kept are
third-party classes that did not originate in this project.

Also update the Checkstyle configuration to not expect any author tags.

Some whitespace fixes are also included where the formatting was
incorrect.

Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>