Remove seemingly superfluous lines of code
This code was added during development of the manual/forced rerouting
feature but appears to be effectively dead. The code before already
takes care of removing any commands that might have already been sent
from the remaining claim. Thus, it should be unnecessary to check if the
remaining claim starts with those commands.

Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>