Ensure proper clearing of order reservations
For vehicles not in integration level TO_BE_UTILIZED, ensure that the
order reservation is cleared so transport orders are not locked in a
reserved state, keeping other vehicles from taking them.

Also ensure that a vehicle has a valid position and is in integration
level TO_BE_UTILIZED before assigning it a reserved transport order.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>