Deprecate report points
Report points require management of special cases in which a vehicle is
not ordered to move to a point, yet still reports that point as reached.
This handling of special cases increases the complexity of the related
code. As vehicle drivers can always report vehicle coordinates for
intermediate positions, report points are a redundant feature that does
not serve any real purpose any more and will be removed for the sake of
simplicity.

Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>