Add support for vehicle envelopes
In an openTCS plant model, envelopes can now be defined for points and
paths a vehicle occupies or traverses. For vehicles, an envelope key can
be defined to indicate which envelopes defined at points and paths
should be considered for the respective vehicle. This way, it is now
possible to prevent vehicles from allocating physical areas intersecting
with areas already allocated by other vehicles.

Also extend the web API specification and implementation with respect 
to the newly added support for vehicle envelopes.

Co-authored-by: Sebastian Bonna <sebastian.bonna@iml.fraunhofer.de>
Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>