Improve handling of failed peripheral jobs
A failed peripheral job (where the completion required flag is set to
true) associated with a transport order now leads to the respective
transport order to be withdrawn. Continuing the transport order in such
cases does not seem reasonable as it is unclear how the outcome of the
peripheral operation influences the vehicle.

Additionally, DefaultVehicleController now properly handles resource
management when peripheral jobs fail.

While here, refactor the code in DefaultVehicleController regarding
handling of peripheral jobs to make it easier to understand and improve
maintainability.

Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>