Improve resource management on vehicle movement
When a vehicle moves to a new position without having been ordered to
move anywhere, allocating and freeing resources is now properly handled.
The availability of the required resources is now checked first. Only
if they are actually available, they are allocated and previously
allocated resources are given up.

Also reduce some code duplication in DefaultScheduler.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>