Document which methods are to be called from the kernel executor thread
Most methods in the *Service classes are supposed to be called only
from the kernel executor thread. This also applies to methods in the
default strategies (i.e. dispatcher, router, scheduler). Where
applicable, this fact is now specified in the javadoc of the
corresponding methods.

With this documentation in place, remove (now superfluous) calls to the
kernel executor inside of default strategy implementations.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>