Allow resources to be avoided in the context of transport orders
When computing a route for a transport order, it is now possible to
define a set of resources (i.e., points, paths or locations) that should
be avoided by vehicles processing the respective transport order. For
this, a property with the key `tcs:resourcesToAvoid` can be set on a
transport order to a comma-separated list of resource names.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>