Avoid creating the same peripheral job twice
This fixes a bug where a peripheral job was created twice if the
vehicle was rerouted before the peripheral job was finished.

When a transport order is updated in the context of rerouting a vehicle,
the future commands and the claim of the vehicle have to be updated as
well. To do this, commands that have already been processed or are still
being processed have to be discarded from the new future commands and
the new claim.

Commands for which a peripheral interaction is pending also belong to 
the category of commands being processed and should therefore be
discarded, too. These commands have not yet been fully processed, as
the associated interaction is still pending, but the resources for them
have already been allocated and ultimately they will be handed over to
the vehicle driver (provided the associated interaction is finished
successfully). If those commands are not discarded then it is possible
that a command with pending interactions is sent to the vehicle twice
if the vehicle is rerouted in the meantime.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>