Make vehicle resource management configurable
The type of how vehicle resources (i.e., paths, points and locations
allocated by vehicles) are managed can now be configured via the
KernelApplicationConfiguration.vehicleResourceManagementType
configuration entry.

With the LENGTH_RESPECTED configuration, resources are released
respecting the length of vehicles. This means that only resources that
are no longer "covered" by a vehicle (according to the length of the
vehicle and the length of the paths behind it) are released. This is the
"classic" way resources were managed before vehicle envelopes were
introduced.

With the LENGTH_IGNORED configuration, resources are released ignoring
the length of vehicles. This means that resources are always released up
to (excluding) a vehicle's current position. This type can be useful
when you primarily want to utilize vehicle envelopes for traffic
management.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>