Add advanced features to ModbusTCPVehicleCommAdapter
The commit introduces advanced features to the ModbusTCPVehicleCommAdapter for the transport of orders, and modifying commands. It also includes the initialization and adding of values to different maps (Map1, map2, mcPool, positionToStation, stationToCommands). Now, basic conversion function that changes a TransportOrder to a list of ModbusCommands is available. Moreover, writeModbusCommands method and two additional functions -getMaxAllowedSpeed and initializeStationMaps are also implemented now.