Refactor ModbusTCPVehicleCommAdapter and add LoadHandlingDevice
This commit primarily refactors the ModbusTCPVehicleCommAdapter, optimizing the processing of movement commands and improving the handling of transport orders. It also introduces the LoadHandlingDevice class to better manage the vehicle's load handling capabilities. Minor modifications in the CustomProcessModelTO class for setting various operations and vehicle speeds have also been done.