Refactor ModbusTCPVehicleCommAdapter for cleaner functions
The code was updated to refactor ModbusTCPVehicleCommAdapter class and to clean up its functions for better maintainability. Comments, unused and commented-out code were removed. Load and Unload commands were included in getLiftCommand() function. The getMaxAllowedSpeed() function was also updated to use the lesser of maxFwdVelocity and maxRevVelocity. Updates also included ensuring the processing of peripheral operations, if available, when creating CMD2 command.