Fix handling of forced rerouting
Previously, forcefully rerouting a vehicle that's waiting for peripheral
jobs to be finished could cause the kernel executor thread to get stuck
in a loop. Therefore, prevent vehicles from being forcefully rerouted in
such situations (and fix the loop-issue in general).

Moreover, prevent the vehicle controller from sending any further
movement commands to the vehicle driver when the vehicle is reported at
an unexpected position. Only after assigning a new transport order to
the vehicle or forcefully rerouting it, the vehicle controller will
continue to send movement commands.

Additionally, fix an issue where a vehicle would not get rerouted
correctly when forcefully rerouting it after it has reported an
unexpected position.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>