Enable and configure DocLint
Enable DocLint and check for possible issues in JavaDoc comments, but
explicitly disable checks for missing documentation comments (or tags).
As we also use Checkstyle, where checks for missing documentation (can
be and) is configured at a finer level, an additional check via DocLint
is unnecessary.

Additionally, configure warnings to be reported as an error, resulting
in the javadoc task to fail if there are any DocLint warnings.

Also fix DocLint issues that would now otherwise prevent the build to
complete successfully.

Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>