Add costs to Route.Step and deprecate Router.getCosts()
Previously, the costs of a route were computed seperately by the router
and then stored in the route. To ensure that the correct costs for the
correct route are computed and to improve performance, the route costs
are now computed through the costs of every single step belonging to the
route. For that, Route.Step now has a member costs.

Additionally, deprecate Router.getCosts() as the equivalent
Router.getRoute() method already provides a way to retrieve the costs of
a route.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>