Add support for transport order types with priorities
For vehicle elements, transport order types (that a vehicle is allowed
to process) can now be assigned a priority.

Also add a new key "BY_ORDER_TYPE_PRIORITY" that can be configured for
the kernel configuration entries
"defaultdispatcher.orderCandidatePriorities" and
"defaultdispatcher.vehicleCandidatePriorities". With this key
configured, a vehicle's acceptable order types are considered when
assigning transport order to it.

Finally, add an endpoint "PUT /vehicles/{NAME}/acceptableOrderTypes",
which allows a vehicle's set of acceptable order types to be modified
during runtime. This endpoint replaces
"PUT /vehicles/{NAME}/allowedOrderTypes", which is now deprecated.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Reviewed-by: Finja Averhaus <finja.averhaus@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>