Extend router API for providing multiple routes
Previously, the router in openTCS could only provide a single route
which was selected according to the router implementation. Here, we make
it possible for a router to multiple different routes, up to a
configurable number. (The default router implementation still only
provides a single route, though.)

Furthermore, a new interface RouteSelector has been added, allowing to
select one of the routes the router has provided according to
implementation-specific criteria. The default implementation selects a
route according to the lowest costs.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>