Properly handle order sequences & dispensable flag
The interaction of order sequences and the dispensable flag of transport
orders is now better defined and improved. A transport order whose
dispensable flag is set and that is also part of an order sequence gets
skipped/aborted when another transport order is added to the sequence.
Additionally, a vehicle may be assigned a new order/sequence in case it
is processing a dispensable transport order, provided that this
transport order is the last order in a sequence that is marked as
complete.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>