Add a watchdog task for transport orders
Add a watchdog task to the kernel which periodically checks the state
of transport orders and publishes a user notification in case a
transport order is considered _idle_ (e.g., a transport order that has
been waiting for a configurable amount of time to be processed by a
vehicle) or in case an order's deadline has expired and the order is
not yet in a final state.
Additionally, a notification is published when a transport order has
reached a final state after being previously considered as idle or
expired.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>