Remove the admin web API
The admin web API has been removed, and its endpoints have been
integrated into the service web API to improve efficiency by running
only a single server.

As before, the v1/kernel endpoint, that is used to shut down the kernel,
is only accessible from the same environment in which the kernel was
started.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>