Fix loading of plant models in file version 7.0.0
With openTCS 7.0.0, loading a plant model with file version 7.0.0 in the
Model Editor  results in a NPE, if there are vehicles for which no
envelope key is defined. This is not intended, as vehicle envelope keys
are optional in file version 7.0.0 and it should be allowed to omit them
in the plant model file.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>